pyproject.toml
[tool.poetry]
name = "hikari-bot"
version = "0.3.9.5"
description = "Nonebot2 HikariBot,支持战舰世界水表查询"
authors = ["benx1n <shirakamikanade@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/benx1n/HikariBot"
repository = "https://github.com/benx1n/HikariBot"
keywords = ["nonebot", "nonebot2", "qqbot", "wows", "wws","bot","stats"]
packages = [
    { include = "hikari_bot", from = "./src/plugins/"}
]

[tool.poetry.dependencies]
python = "^3.8.0"
nonebot2 = { version = "^2.0.0-beta.1", extras = ["fastapi"] }
nonebot-adapter-onebot = "^2.1.0"
nonebot-plugin-htmlrender = "^0.2.0.1"
nonebot-plugin-apscheduler = "^0.2.0"
nonebot_plugin_guild_patch = "^0.2.1"
nonebot-plugin-reboot = "^0.1.3"
httpx = { version = ">=0.24.0", extras = ["http2"] }
Jinja2 = "^3.0.0"
beautifulsoup4 = "^4.11.1"
orjson = "^3.8.11"

[tool.poetry.group.dev.dependencies]
nb-cli = "^0.6.0"
nonebot-plugin-gocqhttp = "^0.6.2"
black = {version = "^23.1a1", allow-prereleases = true}
isort = "^5.12.0"

[tool.nonebot]
#plugins = ["nonebot_plugin_apscheduler", "nonebot_plugin_htmlrender","nonebot_plugin_guild_patch"]
plugins = ["nonebot_plugin_apscheduler", "nonebot_plugin_htmlrender","nonebot_plugin_guild_patch","hikari_bot"]
#plugin_dirs = ["src/plugins"]

[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
# default = true

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
